Use a hash instead of a numeric key for strings translations
Fixes a conflict with WPBakery Visual Composer

If for some reason, the number of strings registered when displaying teh
strings and when saving theme is not the same, the translations are
totally messed. Using a hash of the string as key prevents this.

Side effect: if two strings are identical (but different name), only one
is now displayed.