Fix synchronization of metas when a post/term is saved
In version 2.2.x, the synchronization of metas is handled when a
post/term is saved. Users will expect this to go on in 2.3.x. However
moving the synchronization to add/update/delete is not sufficient.
Indeed if the synchonization of a meta is activated *after* the source
post has been created, updating the it would not copy the metas to the
target (as no meta is created/updated/deleted). It is thus manadatory to
restore the synchronization when the post/term is saved.
Fix https://github.com/polylang/polylang-pro/issues/78