Don't cache translation urls before template_redirect
This fixes a conflict with the combination of Barrel + Barrel Theme
addons + WP Bakery Page Builder which for some reason results in a call
to wp_nav_menu before the main query is known causing the translation
urls to point to the homepage (incl. in the current language).
See conversation 3375.