Fix test for default flag
For some reason, the height and width in the inline style of the html
flag are not outputted in the same order depending on the PHP version.
Thus we can't test eh strign literally. So we use preg_match() to test
the style.
Doing the same for the encoded flag to decrease the line length to
improve readability as it's useless to test base64_encode() itself.