Check that $files is not empty
Now that we initialize modules in PLL_Context -> Polylang::_init(), we pass through PLL_WPML_Config::init() and set PLL_WPML_Config::$files() with an empty array.

When we pass to WPML tests, and do PLL_WPML_Config::instance()->init(), we don't get the file created in wpSetUpBeforeClass(), and the tests fail.